refactor names, separate schemes from init
